requirements.txt
mss==10.0.0
numpy==2.1.3
adbutils==2.8.0
PyWinCtl==0.4.1 
torch-directml==0.2.5.dev240914
torch==2.4.1
torchvision==0.19.1
ultralytics==8.3.57
easyocr==1.7.2
pygetwindow==0.0.9
pyautogui==0.9.54
PyYAML==6.0.2
pywin32==308